fix(env): expand defaults for Pixi
