+++ crates/kernel-env/src/conda.rs
@@ -8,8 +8,8 @@ use anyhow::{anyhow, Context, Result};
 use log::{info, warn};
 use rattler::{default_cache_dir, install::Installer};
 use rattler_conda_types::{
-    Channel, ChannelConfig, GenericVirtualPackage, MatchSpec, ParseMatchSpecOptions,
-    ParseStrictness, Platform, PrefixRecord, Version, VersionSpec,
+    ChannelConfig, GenericVirtualPackage, MatchSpec, ParseMatchSpecOptions, ParseStrictness,
+    Platform, PrefixRecord, Version, VersionSpec,
 };
 use rattler_solve::{resolvo, SolverImpl, SolverTask};
 use serde::{Deserialize, Serialize};
@@ -19,7 +19,10 @@ use std::str::FromStr;
 use std::sync::Arc;
 use std::time::Instant;
 
-use crate::progress::{EnvProgressPhase, ProgressHandler, RattlerReporter};
+use crate::{
+    channels::parse_channels,
+    progress::{EnvProgressPhase, ProgressHandler, RattlerReporter},
+};
 
 /// Conda dependency specification.
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -62,18 +65,6 @@ pub const CONDA_BASE_PACKAGES: &[&str] = &[
     "pyarrow>=14",
 ];
 
-/// The channels represented by Conda's special `defaults` multichannel.
-///
-/// `rattler_conda_types::Channel` resolves ordinary names against
-/// `conda.anaconda.org`; it does not expand Conda's `defaults` alias. Keep the
-/// expansion explicit so an `environment.yml` written by Conda reaches the
-/// official Anaconda repositories rather than a nonexistent community channel.
-pub const ANACONDA_DEFAULT_CHANNELS: &[&str] = &[
-    "https://repo.anaconda.com/pkgs/main",
-    "https://repo.anaconda.com/pkgs/r",
-    "https://repo.anaconda.com/pkgs/msys2",
-];
-
 /// Return [`CONDA_BASE_PACKAGES`] as owned package spec strings for install paths.
 pub fn conda_base_packages() -> Vec<String> {
     CONDA_BASE_PACKAGES
@@ -108,33 +99,6 @@ fn should_enforce_gil_python(deps: &CondaDependencies) -> bool {
         .python
         .as_deref()
         .is_some_and(conda_python_requests_free_threading)
-}
-
-fn parse_channels(
-    declared_channels: &[String],
-    channel_config: &ChannelConfig,
-    platform: Platform,
-) -> Result<Vec<Channel>> {
-    if declared_channels.is_empty() {
-        return Ok(vec![Channel::from_str("conda-forge", channel_config)?]);
-    }
-
-    let mut channels = Vec::new();
-    for declared in declared_channels {
-        if declared == "defaults" {
-            for default_channel in ANACONDA_DEFAULT_CHANNELS {
-                // The msys2 repository is part of Anaconda's Windows defaults;
-                // querying it for Unix platforms only produces missing repodata.
-                if default_channel.ends_with("/msys2") && !platform.is_windows() {
-                    continue;
-                }
-                channels.push(Channel::from_str(default_channel, channel_config)?);
-            }
-        } else {
-            channels.push(Channel::from_str(declared, channel_config)?);
-        }
-    }
-    Ok(channels)
 }
 
 /// Compute the unified env hash for a notebook. Used by the captured-deps
@@ -1418,41 +1382,6 @@ fn find_site_packages(base_path: &std::path::Path) -> Option<String> {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn defaults_expands_to_anaconda_channels_instead_of_community_alias() {
-        let config = ChannelConfig::default_with_root_dir(PathBuf::from("/tmp"));
-        let declared = vec!["defaults".to_string()];
-
-        let unix_channels = parse_channels(&declared, &config, Platform::Linux64).unwrap();
-        let unix_urls = unix_channels
-            .iter()
-            .map(|channel| channel.base_url.as_str())
-            .collect::<Vec<_>>();
-        assert_eq!(
-            unix_urls,
-            vec![
-                "https://repo.anaconda.com/pkgs/main/",
-                "https://repo.anaconda.com/pkgs/r/",
-            ]
-        );
-        assert!(unix_urls
-            .iter()
-            .all(|url| !url.contains("conda.anaconda.org/defaults")));
-
-        let windows_channels = parse_channels(&declared, &config, Platform::Win64).unwrap();
-        assert_eq!(
-            windows_channels
-                .iter()
-                .map(|channel| channel.base_url.as_str())
-                .collect::<Vec<_>>(),
-            vec![
-                "https://repo.anaconda.com/pkgs/main/",
-                "https://repo.anaconda.com/pkgs/r/",
-                "https://repo.anaconda.com/pkgs/msys2/",
-            ]
-        );
-    }
 
     #[test]
     fn test_compute_env_hash_stable() {
